🚚 move main logic of marketplace_indexer in lib.rs
